requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
black==21.5b2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
# dataclasses==0.8
fake-useragent==0.1.11
idna==2.10
importlib-metadata==4.4.0
mypy-extensions==0.4.3
pathspec==0.8.1
python-dotenv==0.17.1
python-telegram-bot==13.5
pytz==2021.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
toml==0.10.2
tornado==6.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.5
zipp==3.4.1
